Fix coverity bug
* Fix unintentional integer overflow bug

+++ src/runtime/GlobalObjectBuiltinTypedArray.cpp
@@ -415,7 +415,7 @@ Value builtinTypedArrayConstructor(ExecutionState& state, Value thisValue, size_
             size_t len = values.size();
             // Perform ? AllocateTypedArrayBuffer(O, len).
             size_t elementSize = obj->elementSize();
-            uint64_t byteLength = len * elementSize;
+            uint64_t byteLength = static_cast<uint64_t>(len) * elementSize;
             ArrayBufferObject* buffer = ArrayBufferObject::allocateArrayBuffer(state, state.context()->globalObject()->arrayBuffer(), byteLength);
             obj->setBuffer(buffer, 0, byteLength, len);
